Forum/requirements.txt
fastapi==0.115.6
pydantic==2.10.5
psycopg==3.2.4
PyJWT==2.10.1
authx==1.4.1
passlib==1.7.4
bcrypt==4.2.1